kvserver,logstore: introduce log StateLoader
Previously the StateLoader accessed both log and state machine state. This
commit moves most of the log-specific accesstors to the logstore package.

Release note: None
